+++ src/main/java/com/example/demo/service/OrderService.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Service;
 
 @Service
 public interface OrderService {
-    Order getOrderbyId(long orderId);
-    Order getOrderbyUserId(long userId);
+    Order getOrderbyId(int orderId);
+    Order getOrderbyUserId(int userId);
     Order addOrder(Order order);
 }
